Error que no pudo ser reparado. Algo que ver con el versionamiento.

+++ src/main/java/com/manytoone/demo/persistencia/entity/Producto.java
@@ -13,7 +13,7 @@ public class Producto {
     private String precio;
     private String imagen;
     @Column(name = "id_categoria")
-    private Integer idCategoria;
+    private Long idCategoria;
     @ManyToOne
     @JoinColumn(name = "id_categoria", insertable = false, updatable = false)
     private Categoria categoria;
@@ -50,11 +50,11 @@ public class Producto {
         this.imagen = imagen;
     }
 
-    public Integer getIdCategoria() {
+    public Long getIdCategoria() {
         return idCategoria;
     }
 
-    public void setIdCategoria(Integer idCategoria) {
+    public void setIdCategoria(Long idCategoria) {
         this.idCategoria = idCategoria;
     }
